requirements.txt
playwright
pytest
pytest-playwright
pytest-axe
allure-pytest
python-dotenv
pytest-xdist
faker